+++ spring-cloud-starter-tencent-polaris-circuitbreaker/src/main/java/com/tencent/cloud/polaris/circuitbreaker/config/PolarisCircuitBreakerFeignClientAutoConfiguration.java
@@ -18,14 +18,20 @@
 package com.tencent.cloud.polaris.circuitbreaker.config;
 
 import com.tencent.cloud.polaris.circuitbreaker.feign.PolarisCircuitBreakerNameResolver;
+import com.tencent.cloud.polaris.circuitbreaker.feign.PolarisFeignCircuitBreaker;
+import com.tencent.cloud.polaris.circuitbreaker.feign.PolarisFeignCircuitBreakerTargeter;
 import feign.Feign;
 
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.cloud.client.circuitbreaker.CircuitBreakerFactory;
 import org.springframework.cloud.openfeign.CircuitBreakerNameResolver;
 import org.springframework.cloud.openfeign.FeignClientFactoryBean;
+import org.springframework.cloud.openfeign.Targeter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Scope;
 
 /**
  * PolarisCircuitBreakerFeignClientAutoConfiguration.
@@ -41,6 +47,21 @@ public class PolarisCircuitBreakerFeignClientAutoConfiguration {
 	@ConditionalOnMissingBean(CircuitBreakerNameResolver.class)
 	public CircuitBreakerNameResolver polarisCircuitBreakerNameResolver() {
 		return new PolarisCircuitBreakerNameResolver();
+	}
+
+	@Bean
+	@ConditionalOnBean(CircuitBreakerFactory.class)
+	@ConditionalOnMissingBean(Targeter.class)
+	public Targeter polarisFeignCircuitBreakerTargeter(CircuitBreakerFactory circuitBreakerFactory, CircuitBreakerNameResolver circuitBreakerNameResolver) {
+		return new PolarisFeignCircuitBreakerTargeter(circuitBreakerFactory, circuitBreakerNameResolver);
+	}
+
+	@Bean
+	@Scope("prototype")
+	@ConditionalOnBean(CircuitBreakerFactory.class)
+	@ConditionalOnMissingBean(Feign.Builder.class)
+	public Feign.Builder circuitBreakerFeignBuilder() {
+		return PolarisFeignCircuitBreaker.builder();
 	}
 
 }